Line.hpp: Add new line component to rendering

+++ src/Geometry/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.hpp"
 
-Point::Point(int x, int y, int z) {
+Point::Point(float x, float y, float z) {
   this->x = x;
   this->y = y;
   this->z = z;
